LUN-3: Move Interfaces from professionSubcategory.ts
