--- src/interfaces/professionSubcategory.ts
@@ -1,12 +1,0 @@
-export interface IProfessionSubcategoryScreenProps {
-  route: any
-  navigation: any
-}
-
-export interface IProfessionSubcategoryProps {
-  id: number
-  title: string
-  description: string
-  icon: string
-  total_documents: number
-}
